dev: moved shipping tour into guided tours folder
